serializer/file_test.go
package serializer

import (
	"github.com/golang/protobuf/proto"
	"github.com/igilgyrg/gRPC-pet/app"
	"github.com/igilgyrg/gRPC-pet/pb"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
	"testing"
)

func TestWriteProtobufToBinaryFile(t *testing.T) {
	t.Parallel()

	binaryFile := "../tmp/laptop.bin"
	jsonFile := "../tmp/laptop.json"

	laptop1 := app.NewLaptop()
	err := WriteProtobufToBinaryFile(laptop1, binaryFile)
	require.NoError(t, err)

	laptop2 := &pb.Laptop{}
	err = ReadProtobufFromBinaryFile(binaryFile, laptop2)
	require.NoError(t, err)
	assert.Equal(t, laptop1.Id, laptop2.Id)
	require.True(t, proto.Equal(laptop1, laptop2))

	err = WriteProtobufToJSONFile(laptop2, jsonFile)
	require.NoError(t, err)
}
